+++ pyobjc-framework-CoreMotion/PyObjCTest/test_cmfalldetectionevent.py
@@ -1,0 +1,10 @@
+import CoreMotion
+from PyObjCTools.TestSupport import TestCase
+
+
+class TestCMFallDetectionEvent(TestCase):
+    def test_constants(self):
+       self.assertEqual(CoreMotion.CMFallDetectionEventUserResolutionConfirmed, 0)
+       self.assertEqual(CoreMotion.CMFallDetectionEventUserResolutionDismissed, 1)
+       self.assertEqual(CoreMotion.CMFallDetectionEventUserResolutionRejected, 2)
+       self.assertEqual(CoreMotion.CMFallDetectionEventUserResolutionUnresponsive, 3)