+++ pyobjc-framework-NetworkExtension/PyObjCTest/test_neappproxyflow.py
@@ -25,3 +25,9 @@ class TestNEAppProxyFlow(TestCase):
             1,
             b"v@",
         )
+
+    @min_os_level("10.16.1")
+    def testMethods10_16_1(self):
+        self.assertResultIsBOOL(
+            NetworkExtension.NEAppProxyFlow.isBound
+        )